First hacks at running compute split-brain
By setting COMPUTE_BACKEND_PORT to something other than COMPUTE_HTTP_PORT and COMPUTE_HTTPS_PORT, compute will start itself in proxy mode in one process, and then spawn a second process to handle the actual compute computations. This makes it so that the front end can do the stashing and logging even if the back end crashes during the computation. Hoping that this will make it so that we can make sure to always get the input that causes Rhino to crash. With Rhino running in-process, it's highly likeyl that the crash will happen before the input is logged; causing us to lose all of it.